Created file for functions
Created a file for the backend functions and moved them into it.

The current logic is more or less proof of concept so need to implement it from scratch again
