+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "core",
-	"version": "1.0.0",
+	"version": "1.0.1",
 	"scripts": {
 		"install": "concurrently \"cd backend && npm install\" \"cd frontend && npm install\"",
 		"dev": "concurrently \"cd backend && npm run dev\" \"cd frontend && npm run dev\"",
